+++ src/features/articles/hooks/index.ts
@@ -1,5 +1,5 @@
 export { useLike } from './useLike/useLike';
 export { useStock } from './useStock/useStock';
-export { useCommentForm } from './useCommentForm';
-export { useCommentItem } from './useCommentItem';
-export { useArticleDelete } from './useArticleDelete';
+export { useCommentForm } from './useCommentForm/useCommentForm';
+export { useCommentItem } from './useCommentItem/useCommentItem';
+export { useArticleDelete } from './useArticleDelete/useArticleDelete';
